Refactor large routes into separate blueprints
