+++ all_routes.py
@@ -543,398 +543,3 @@ def view_hierarchy():
 
 
 
-@routes_bp.route('/api/stats')
-def get_stats():
-    """Return accurate counts of diagrams and rules"""
-    try:
-        diagrams_dir = current_app.config.get('DIAGRAMS_FOLDER', './diagrams')
-        diagrams_count = 0
-        rules_count = 0
-
-        if os.path.exists(diagrams_dir):
-            # Count all .mmd files as diagrams
-            diagrams_count = sum(
-                1 for root, _, files in os.walk(diagrams_dir)
-                for file in files
-                if file.endswith('.mmd')
-            )
-
-            # Count rules from all JSON files
-            for root, _, files in os.walk(diagrams_dir):
-                for file in files:
-                    if file.endswith('.json'):
-                        with open(os.path.join(root, file), 'r') as f:
-                            data = json.load(f)
-                            if isinstance(data, list):
-                                rules_count += len(data)
-                            elif isinstance(data, dict):
-                                rules_count += len(data.get('rules', []))
-
-        return jsonify({
-            'diagramsProcessed': diagrams_count,
-            'rulesExtracted': rules_count,
-            'status': 'success'
-        })
-
-    except Exception as e:
-        current_app.logger.error(f"Stats error: {str(e)}")
-        return jsonify({
-            'status': 'error',
-            'message': 'Could not calculate statistics'
-        }), 500
-
-@routes_bp.route('/api/advanced_stats')
-def get_advanced_stats():
-    """Return business metrics tracked in a JSON file."""
-    try:
-        json_path = os.path.join(current_app.root_path, 'data', 'advanced_stats.json')
-        if not os.path.exists(json_path):
-            # Create a default structure if the file doesn't exist
-            stats = {
-                "rules_by_status": {"active": 0, "draft": 0, "deprecated": 0},
-                "recent_changes": 0,
-                "validation_errors": 0,
-                "active_domain": "N/A"
-            }
-            # Ensure the directory exists before writing
-            os.makedirs(os.path.dirname(json_path), exist_ok=True)
-            with open(json_path, 'w', encoding='utf-8') as f:
-                json.dump(stats, f)
-        else:
-            with open(json_path, 'r', encoding='utf-8') as f:
-                stats = json.load(f)
-        return jsonify(stats)
-    except Exception as e:
-        current_app.logger.error(f"Advanced stats error: {str(e)}")
-        return jsonify({'error': str(e)}), 500
-
-@routes_bp.route('/api/analytics/activity', methods=['GET'])
-def get_activity():
-    Config.ensure_data_dir()
-    try:
-        return send_file(
-            Config.ACTIVITY_LOG,
-            mimetype='application/json',
-            as_attachment=False  # Set to True if you want it to download
-        )
-    except FileNotFoundError:
-        return jsonify({"error": "Activity log not found"}), 404
-
-
-@routes_bp.route('/api/activity_stats')
-def get_activity_stats():
-    """Return comprehensive activity statistics with accurate counting"""
-    try:
-        Config.ensure_data_dir()
-
-        # Initialize stats with proper defaults
-        stats = {
-            "status_counts": {"active": 0, "draft": 0, "deprecated": 0},
-            "recent_activity": {"total": 0, "creations": 0, "updates": 0, "deletions": 0},
-            "top_contributor": {"user": "None", "actions": 0}
-        }
-
-        try:
-            with open(Config.ACTIVITY_LOG, 'r') as f:
-                data = json.load(f)
-
-                # 1. Count rules by status
-                for rule in data.get('rules', {}).values():
-                    status = rule.get('status', 'draft').lower()
-                    if status in stats['status_counts']:
-                        stats['status_counts'][status] += 1
-
-                # 2. Process activity log
-                user_actions = defaultdict(int)
-                week_ago = datetime.utcnow() - timedelta(days=7)
-
-                for entry in data.get('activity_log', []):
-                    try:
-                        # Count user actions
-                        user = entry.get('user', 'anonymous')
-                        user_actions[user] += 1
-
-                        # Categorize activity
-                        action = entry.get('action', '').lower()
-                        if 'create' in action:
-                            stats['recent_activity']['creations'] += 1
-                        elif 'delete' in action:
-                            stats['recent_activity']['deletions'] += 1
-                        elif 'update' in action:
-                            stats['recent_activity']['updates'] += 1
-
-                        # Count if recent
-                        entry_time = datetime.fromisoformat(entry['timestamp'])
-                        if entry_time >= week_ago:
-                            stats['recent_activity']['total'] += 1
-
-                    except Exception as e:
-                        current_app.logger.warning(f"Skipping invalid entry: {e}")
-                        continue
-
-                # 3. Determine top contributor
-                if user_actions:
-                    top_user, top_actions = max(user_actions.items(), key=lambda x: x[1])
-                    stats['top_contributor'] = {
-                        "user": top_user,
-                        "actions": top_actions
-                    }
-
-        except (FileNotFoundError, json.JSONDecodeError) as e:
-            current_app.logger.error(f"Activity log error: {e}")
-
-        return jsonify(stats)
-
-    except Exception as e:
-        current_app.logger.error(f"Stats generation failed: {e}")
-        return jsonify({"error": str(e)}), 500
-
-@routes_bp.route('/api/rules/<rule_id>', methods=['PUT'])
-def update_rule(rule_id):
-    data = request.get_json()
-
-    try:
-        # Your existing update logic here...
-
-        # Log the activity
-        log_activity(
-            action="update",
-            rule_id=rule_id,
-            user=get_current_user(),  # Implement your user auth
-            details=f"Updated rule {rule_id}"
-        )
-
-        return jsonify({"success": True})
-
-    except Exception as e:
-        current_app.logger.error(f"Rule update error: {str(e)}")
-        return jsonify({"error": str(e)}), 500
-
-@routes_bp.route('/test/log_activity')
-def test_log_activity():
-    """Test endpoint to verify activity logging"""
-    try:
-        log_activity(
-            action="test",
-            rule_id="test_rule_123",
-            user="tester",
-            details="Test activity logging"
-        )
-        return jsonify({"status": "success", "message": "Activity logged"})
-    except Exception as e:
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-@routes_bp.route('/test/view_activity_log')
-def test_view_activity_log():
-    """Test endpoint to view raw activity log"""
-    try:
-        Config.ensure_data_dir()
-        with open(Config.ACTIVITY_LOG, 'r') as f:
-            return jsonify(json.load(f))
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-@routes_bp.route('/force_log_activity')
-def force_log_activity():
-    """Force log a test activity"""
-    try:
-        log_activity(
-            action="test_action",
-            rule_id="test_rule_123",
-            user="test_user",
-            details="Test activity entry"
-        )
-        return jsonify({"status": "success", "message": "Test activity logged"})
-    except Exception as e:
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-
-# ================
-# USER ROUTES
-# ================
-
-@routes_bp.route('/profile')
-def user_profile():
-    """User profile page with activity and statistics"""
-    try:
-        # Get user data (in a real app, this would come from your database)
-        user_data = {
-            "username": "johndoe",
-            "email": "john.doe@example.com",
-            "avatar": url_for('static', filename='images/default-avatar.png'),
-            "member_since": "2023-01-15",
-            "last_login": datetime.now().strftime("%Y-%m-%d %H:%M"),
-            "role": "Administrator"
-        }
-
-        # Get user activity stats
-        activity_stats = {
-            "rules_created": 142,
-            "diagrams_uploaded": 28,
-            "api_calls": 1240,
-            "recent_activity": [
-                {"action": "Uploaded rule set", "timestamp": "2023-06-15 14:30", "details": "Fraud Detection v2.1"},
-                {"action": "Extracted rules", "timestamp": "2023-06-14 09:15", "details": "From production dataset"},
-                {"action": "API Test", "timestamp": "2023-06-12 16:45", "details": "/rules/validate endpoint"}
-            ]
-        }
-
-        return render_template(
-            'profile.html',
-            user=user_data,
-            stats=activity_stats,
-            help_available=True
-        )
-
-    except Exception as e:
-        current_app.logger.error(f"Profile error: {str(e)}")
-        abort(500, "Error loading profile page")
-
-
-@routes_bp.route('/settings', methods=['GET', 'POST'])
-def user_settings():
-    """User settings page with form handling"""
-    try:
-        # Dummy user data (replace with real user lookup in production)
-        user_data = {
-            "display_name": "John Doe",
-            "email": "john.doe@example.com",
-            "avatar_url": url_for('static', filename='images/default-avatar.png')
-        }
-
-        # Default settings data
-        settings_data = {
-            "theme": "dark",
-            "timezone": "UTC",
-            "email_notifications": True,
-            "push_notifications": False,
-            "inapp_notifications": True,
-            "language": "en",
-            "experimental_features": False,
-            "font_size": "normal",
-            "contrast": "normal",
-            "motion": "normal"
-        }
-
-        # Handle form submission
-        if request.method == 'POST':
-            settings_data.update({
-                "theme": request.form.get('theme', 'dark'),
-                "timezone": request.form.get('timezone', 'UTC'),
-                "email_notifications": 'email_notifications' in request.form,
-                "push_notifications": 'push_notifications' in request.form,
-                "inapp_notifications": 'inapp_notifications' in request.form,
-                "language": request.form.get('language', 'en'),
-                "experimental_features": 'experimental_features' in request.form,
-                "font_size": request.form.get('font_size', 'normal'),
-                "contrast": request.form.get('contrast', 'normal'),
-                "motion": request.form.get('motion', 'normal')
-            })
-            # Save logic here
-            flash('Settings updated successfully!', 'success')
-
-        security_data = {
-            "last_password_change": "2023-05-10",
-            "two_factor_enabled": False,
-            "active_sessions": [
-                {"ip": "192.168.1.100", "browser": "Chrome", "location": "New York", "last_active": "10 minutes ago"}
-            ]
-        }
-
-        return render_template(
-            'settings.html',
-            user=user_data,
-            settings=settings_data,
-            security=security_data,
-            help_available=True
-        )
-
-    except Exception as e:
-        current_app.logger.error(f"Settings error: {str(e)}")
-        abort(500, "Error loading settings page")
-
-
-@routes_bp.route('/api/user/profile', methods=['GET'])
-def get_user_profile():
-    """API endpoint for user profile data"""
-    try:
-        # In a real app, this would come from your database
-        profile_data = {
-            "username": "johndoe",
-            "email": "john.doe@example.com",
-            "avatar_url": url_for('static', filename='images/default-avatar.png'),
-            "member_since": "2023-01-15",
-            "role": "Administrator",
-            "stats": {
-                "rules_created": 142,
-                "diagrams_uploaded": 28,
-                "api_calls": 1240
-            }
-        }
-        return jsonify(profile_data)
-    except Exception as e:
-        current_app.logger.error(f"Profile API error: {str(e)}")
-        return jsonify({"error": str(e)}), 500
-
-
-@routes_bp.route('/api/user/activity', methods=['GET'])
-def get_user_activity():
-    """API endpoint for user activity data"""
-    try:
-        # Example activity data - in a real app, this would come from your database
-        activity_data = [
-            {
-                "id": 1,
-                "action": "rule_upload",
-                "description": "Uploaded Fraud Detection v2.1 rules",
-                "timestamp": "2023-06-15T14:30:00Z",
-                "details": {
-                    "file": "fraud_detection_v2.1.json",
-                    "rule_count": 24
-                }
-            },
-            {
-                "id": 2,
-                "action": "rule_extraction",
-                "description": "Extracted rules from production dataset",
-                "timestamp": "2023-06-14T09:15:00Z",
-                "details": {
-                    "source": "production_db",
-                    "rule_count": 18
-                }
-            }
-        ]
-        return jsonify(activity_data)
-    except Exception as e:
-        current_app.logger.error(f"Activity API error: {str(e)}")
-        return jsonify({"error": str(e)}), 500
-
-
-@routes_bp.route('/api/user/settings', methods=['GET', 'PUT'])
-def handle_user_settings():
-    """API endpoint for user settings"""
-    try:
-        if request.method == 'GET':
-            # Return current settings
-            settings = {
-                "theme": "dark",
-                "timezone": "UTC",
-                "notifications": {
-                    "email": True,
-                    "push": False
-                },
-                "language": "en"
-            }
-            return jsonify(settings)
-
-        elif request.method == 'PUT':
-            # Update settings
-            new_settings = request.get_json()
-            current_app.logger.info(f"Updating settings: {new_settings}")
-
-            # In a real app, you would save these to the database
-            return jsonify({"status": "success", "message": "Settings updated"})
-
-    except Exception as e:
-        current_app.logger.error(f"Settings API error: {str(e)}")
-        return jsonify({"error": str(e)}), 500
